refactor(assembly): extract shared rezipWithMimetypeFirst helper
- rezipWithMimetypeFirst was byte-identical in flattenFooterFields.ts and
  renameMasterPageStyles.ts; move it to src/server/xml/rezipWithMimetypeFirst.ts
  and import from both call sites
- behavior-preserving: existing flattenFooterFields.test.ts and
  renameMasterPageStyles.test.ts pass unchanged
